chore: Moved all elements tied to server into core/server.
