+++ packages/architectura/src/core/execution-context/execution-context.service.mts
@@ -1,13 +1,15 @@
 
 
-import type { RichClientRequest } from "../../core/rich-client-request.mjs";
 
-import type { RichServerResponse } from '../../core/rich-server-response.mjs';
 
 
 import type { ExecutionContextInstantiationInterface } from "../../definition/interface/execution-context-instantiation.interface.mjs";
 
-import type { SessionService } from "../../index.mjs";
+
+
+import type { RichClientRequest } from "../server/rich-client-request.mjs";
+
+import type { RichServerResponse } from "../server/rich-server-response.mjs";
 
 
 class ExecutionContextService
