+++ packages/architectura/src/core/server/rich-client-request.mts
@@ -4,9 +4,9 @@ import { parse as parseQuery } from "node:querystring";
 
 import { TypeAssertion } from "@vitruvius-labs/ts-predicate";
 
-import { ContentTypeEnum } from "../definition/enum/content-type.enum.mjs";
+import { ContentTypeEnum } from "../../definition/enum/content-type.enum.mjs";
 
-import { LoggerProxy } from "../service/_index.mjs";
+import { LoggerProxy } from "../../service/_index.mjs";
 
 import type { IncomingHttpHeaders } from "node:http";
